backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
boto3
pdf2image
pytesseract
pypdf2
pdfplumber
nltk
spacy
regex
python-multipart
openai
transformers
langchain
tabulate
rich
